Drop video file and rename VidePlayer to Video
