+++ apps/journeys/src/components/blocks/Video/Video.stories.tsx
@@ -1,5 +1,5 @@
 import { Story, Meta } from '@storybook/react'
-import { Video } from './'
+import { Video } from '.'
 import { TreeBlock } from '../../../libs/transformer/transformer'
 import { journeysConfig } from '../../../libs/storybook/decorators'
 import { GetJourney_journey_blocks_VideoBlock as VideoBlock } from '../../../../__generated__/GetJourney'
