+++ apps/journeys/src/components/blocks/Video/index.ts
@@ -1,1 +1,1 @@
-export { Video } from './Video'
+export { Video } from "./Video";
